// src/api/types/SignupFormStylesItemSelector.ts
//  This file was auto-generated from our API Definition.

/** A string that identifies the element selector. */
export const SignupFormStylesItemSelector = {
    PageBackground: "page_background",
    PageHeader: "page_header",
    PageOuterWrapper: "page_outer_wrapper",
    BodyBackground: "body_background",
    BodyLinkStyle: "body_link_style",
    FormsButtons: "forms_buttons",
    FormsButtonsHovered: "forms_buttons_hovered",
    FormsFieldLabel: "forms_field_label",
    FormsFieldText: "forms_field_text",
    FormsRequired: "forms_required",
    FormsRequiredLegend: "forms_required_legend",
    FormsHelpText: "forms_help_text",
    FormsErrors: "forms_errors",
    MonkeyRewardsBadge: "monkey_rewards_badge",
} as const;
export type SignupFormStylesItemSelector =
    (typeof SignupFormStylesItemSelector)[keyof typeof SignupFormStylesItemSelector];